chromium: Unconditionally install files we always have or build.
Now that we have only one recipe, we can drop a few checks and always assume
that files in files/ or binaries we build are present.

While here, separate directory creation from file installation so that it is
easier to know which directories we are really creating in the installation.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>