chromium: Pass additional ARM-related flags to gyp.
In addition to arm_float_abi, we also need to specify some additional
flags:
* arm_arch
* arm_fpu
* arm_tune
* arm_version (we really need this one for ARMv6 builds, as it defaults to
  "7").

The arch-specific include.gypi files were setting the 'armv7' gyp variable,
even though it has had no effect since 2013 (see Chromium bug 234135), so
we can drop it while here.

If not specified, GN will default to some generic values such as "armv7-a".

These values are then passed to the compiler via -march, -mfpu and -mtune.
Since our toolchain also often passes those arguments to the compiler with
different values due to TUNE_CCARGS, we can end up with conflicting options
being passed to the compiler, with Chromium's being passed later and thus
having precedence.

The easiest way to deduce the values that the toolchain passes are by
parsing TUNE_CCARGS and setting the GN variables accordingly. This means
these arguments will be repeated in the command line, but this should not be
an issue.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>